Enhance delete confirmation logic in CompaniesComponent to handle user response and update dialog state

+++ src/app/pages/companies/companies.component.ts
@@ -30,7 +30,12 @@ export class CompaniesComponent {
     }
   }
 
-  onDelete() {
+  onDelete(confirm: boolean) {
+    if (!confirm) {
+      this.openDialogDelete = false;
+      this.selectedCompany = null;
+      return;
+    }
     this.companiesService.delete(this.selectedCompany?.id!).pipe(catchError(error => {
       toast.Notify.failure('No se pudo eliminar la empresa');
       return error
